Move env.warn_and_replace() to SphinxInput class
