+++ sphinx/io.py
@@ -8,6 +8,8 @@
     :copyright: Copyright 2007-2017 by the Sphinx team, see AUTHORS.
     :license: BSD, see LICENSE for details.
 """
+import codecs
+
 from docutils.io import FileInput, NullOutput
 from docutils.core import Publisher
 from docutils.readers import standalone
@@ -25,6 +27,7 @@ from sphinx.transforms.compact_bullet_list import RefOnlyBulletListTransform
 from sphinx.transforms.i18n import (
     PreserveTranslatableMessages, Locale, RemoveTranslatableInline,
 )
+from sphinx.util import logging
 from sphinx.util import import_object, split_docinfo
 from sphinx.util.docutils import LoggingReporter
 
@@ -38,6 +41,9 @@ if False:
     from sphinx.application import Sphinx  # NOQA
     from sphinx.builders import Builder  # NOQA
     from sphinx.environment import BuildEnvironment  # NOQA
+
+
+logger = logging.getLogger(__name__)
 
 
 class SphinxBaseReader(standalone.Reader):
@@ -154,6 +160,10 @@ class SphinxFileInput(FileInput):
         # type: (Sphinx, BuildEnvironment, Any, Any) -> None
         self.app = app
         self.env = env
+
+        # set up error handler
+        codecs.register_error('sphinx', self.warn_and_replace)  # type: ignore
+
         kwds['error_handler'] = 'sphinx'  # py3: handle error on open.
         FileInput.__init__(self, *args, **kwds)
 
@@ -186,6 +196,21 @@ class SphinxFileInput(FileInput):
             if self.env.config.rst_prolog:
                 data = self.env.config.rst_prolog + '\n' + data
         return docinfo + data
+
+    def warn_and_replace(self, error):
+        # type: (Any) -> Tuple
+        """Custom decoding error handler that warns and replaces."""
+        linestart = error.object.rfind(b'\n', 0, error.start)
+        lineend = error.object.find(b'\n', error.start)
+        if lineend == -1:
+            lineend = len(error.object)
+        lineno = error.object.count(b'\n', 0, error.start) + 1
+        logger.warning('undecodable source characters, replacing with "?": %r',
+                       (error.object[linestart + 1:error.start] + b'>>>' +
+                        error.object[error.start:error.end] + b'<<<' +
+                        error.object[error.end:lineend]),
+                       location=(self.env.docname, lineno))
+        return (u'?', error.end)
 
 
 def read_doc(app, env, filename):
